Add a bunch of scripts
Should have committed a lot sooner.
Got into the flow Lol. Anyways I
made functions for the player input
and the computer random and tested
them both to an acceptable level.
Am now working on the code for each
round. Have most of it in place, just
need to figure out and code a way to
keep score. I have an idea and am
gonna fiddle with it a bit till i get
it to work the way i want.