Finish game
Got the score to work correctly and clear
after each game. And declares the final
winner with both scores shown. Choices
of each round are in the console. Prompts
and alerts dictate each round and winner
declaration. Inputs are tested and if
incorrect a random choice is taken.
everything tested multiple times and
appears to work fine. Whew! first
project down!